shootout-reports/reports/requirements.txt
cycler==0.10.0
matplotlib==1.5.3
numpy==1.11.3
pandas==0.19.2
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
wheel==0.24.0
PyYAML==3.12
jupyter==1.0.0
scipy==0.18.0
